+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Add Phoenix integration w/ JaSerializer.PhoenixView.
   * Infer type from module name.
   * Add `attributes/2` callback w/ default implementation based on `attributes/1` macro.
+  * Add JaSerializer.ContentTypeNegotiation plug.
 * Deprecations:
   * Remove `serialize` macro in favor of `type/0` callback.